+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -445,10 +445,6 @@ instance Crypto c => Arbitrary (ShelleyDelegCert c) where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
-instance Crypto c => Arbitrary (Delegation c) where
-  arbitrary = genericArbitraryU
-  shrink = genericShrink
-
 instance Crypto c => Arbitrary (GenesisDelegCert c) where
   arbitrary = genericArbitraryU
   shrink = genericShrink